trailer: don't use 'struct arg_item' for storing config
The '--trailer' options given to 'git interpret-trailers' are store in
the suitably named 'struct arg_item'.

The configuration done in 'trailer.<name>.xyz' was also stored in that
struct. Even though it only needs the "conf_info" part of it.

This commit creates a separate struct for conf_info_item

No functional change intended.

Signed-off-by: Anders Waldenborg <anders@0x63.nu>